Productionise ImagePlacer component.
This component is a polished reusable solution for allowing the user to place an image within a view area, and export visible image.

The component supports touch, svg, serialisation, and provides a consistent panning and zooming experience regardless of the image size (smaller or larger than container) when using constraints and zoom fitting.