Refactor renderer item type into a disjoint union to enable flow types
This covers all of the inbuilt item types and they are now properly
being type checked.
Disjoint union types do not work with a non-distinct or mixed type for
the 'type' identifier prop. We may have to change the API to allow these
new types to work with custom and inline components.