Extend examples iframes to support loading examples from other paths
We want to co-locate some docs-specific examples to the docs directory
so they do not appear in the examples list and so they are colocated
with the docs better. We already load the docs in our site data which is
included in the examples bundle so this is the only change required.