FS-3310 Fix handling of duplicate users in mention typeahead
The 'title' field set on items (TypeaheadItem) provided to the generic
typeahead plugin is used as the 'key' attribute when rendering them.
The problem is that the mention plugin is using the user's full name as value
for the 'title' field. This can cause the typeahead to render duplicate items
when results contain more than 1 user with the same full name.

As a quick fix, we are instead passing the user's id which is expected to be
unique. A better fix would probably to introduce a mandatory 'id' field to
TypeaheadItem to ensure integrators provide a unique value to be used as the
key attribute when rendering. However it would introduce a breaking change for
consumers of the Editor.