FS-3327 Remove client-side caching from mentions
Remove dependency on js-search and caching of remote results on the client side.
With the introduction of URS-backed mention results, the gap in the ranking
logic between client side and server side is becoming more and more noticeable.
In particular, it creates an effect of results jumping around as the results
received from the remote service replace the ones coming from the local index.