FS-3327 Remove unused getUsersInContext from MentionResource
getUsersInContext was initially introduced to rerank users to be rendered so
that provided "users in context" are prioritised over other ones. The main
(if only) use case was for prioritising people that are members of a public
room in Stride. This ranking logic is not used in other products (Confluence)
at the moment and will actually be handled on the service side (URS).