FS-3500 Refactor editor's mention plugin tests
A few changes happening here to improve speed and stability of the test suite:
- Replace usage of hardcoded constant sleep() calls with a promise resolving
when mention results are returned. This brings each test from ~250ms to ~50ms,
- Use MockMentionResource without any random wait times to make the tests
deterministic since we're not interested here in testing our-of-order requests
- Remove unnecessary stuff (e.g. contextIdentifierProvider) and refactor tests
to move test setup in a beforeEach, making each test focus on its purpose.