Refactor TokenParser type to support context
Refactor all tokens to use TokenParser type in order to use a common
type in all token modules.

Refactor TokenParser to receive an object instead of a list of
parameters. This way the order doesn't matter anymore and the call is
always explicit (avoid the overhead to check parameter name and order).

Refactor TokenParser to receive a context object which is an immutable
shared configuration to all tokens.