Re-render Comment components if the `portal` prop changes
In implementation, the ref that needs to be supplied to a portal is often not ready when the component
initially mounts (e.g., if the "portal" sits alongside the Conversation or Comment using it), so the prop
passed to a Conversation or Comment will briefly be `null` before updating to the DOM node. The previous
implementation of `shouldComponentUpdate` would block a necessary re-render when the `portal` value is finally
set.